% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{c_datatable}
\alias{c_datatable}
\title{Create an HTML table widget using the DataTables library with Chinese menu}
\usage{
c_datatable(...)
}
\arguments{
\item{pageLength}{the length of page}

\item{lengthMenu}{the menu of length}

\item{buttons}{buttons shows}
}
\description{
Create an HTML table widget using the DataTables library with Chinese menu
}
