% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_col_defs}
\alias{create_col_defs}
\title{create column definition for columns}
\usage{
create_col_defs(df, pattern, col_def)
}
\arguments{
\item{df}{dataframe}

\item{pattern}{pattern of colnames}

\item{col_def}{column definition}
}
\value{
A list of column definition
}
\description{
create column definition for columns
}
