% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{download_image_url}
\alias{download_image_url}
\title{download image of the image column}
\usage{
download_image_url(
  df,
  df_image_url_col_name = "图片",
  df_image_title_col_name = "SKU",
  image_save_path = getOption("sku_image_save_path", "~/images")
)
}
\arguments{
\item{df}{a tbl or dataframe}

\item{df_image_url_col_name}{Column name of image url}

\item{df_image_title_col_name}{Column name of image title}

\item{image_save_path}{The path images saved to.}
}
\value{
df
}
\description{
download image of the image column
}
