% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{loginUI}
\alias{loginUI}
\title{login UI module}
\usage{
loginUI(
  id,
  title = "请登录",
  user_title = "用户名",
  pass_title = "密码",
  login_title = "登录",
  error_message = "用户名或密码错误"
)
}
\arguments{
\item{id}{Shiny id}

\item{title}{header title for the login panel}

\item{user_title}{label for the user name text input}

\item{pass_title}{label for the password text input}

\item{login_title}{label for the login button}

\item{error_message}{error messsage when failed.}
}
\value{
Shiny UI
}
\description{
Call via \code{loginUI("your_id")}
}
