% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{set_func_defaults}
\alias{set_func_defaults}
\title{Set defaults for function}
\usage{
set_func_defaults(func, recursive = FALSE, ...)
}
\arguments{
\item{func}{Function}

\item{recursive}{Recursively merge defaults with new arguments or not}

\item{...}{Default arguments}
}
\description{
Set defaults for function
}
