% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{write_excel_image}
\alias{write_excel_image}
\title{Write dataframe with image url to excel file with image column}
\usage{
write_excel_image(
  df,
  file = tempfile(fileext = ".xlsx"),
  df_image_url_col_name = "图片",
  df_image_title_col_name = "SKU",
  image_save_path = getOption("sku_image_save_path", "~/images"),
  widths = 27,
  heights = 144,
  width = 5,
  height = 5,
  units = "cm",
  dpi = 300
)
}
\arguments{
\item{df}{a tbl or dataframe}

\item{file}{A character string naming an xlsx file}

\item{df_image_url_col_name}{Column name of image url}

\item{df_image_title_col_name}{Column name of image title}

\item{image_save_path}{The path images saved to.}

\item{widths}{widths to set cols to specified in Excel column width units or "auto" for automatic sizing. The widths argument is
recycled to the length of cols.}

\item{heights}{Heights to set rows to specified in Excel column height units.}

\item{width}{Width of figure.}

\item{height}{Height of figure.}

\item{units}{Units of width and height. Can be "in", "cm" or "px"}

\item{dpi}{Image resolution used for conversion between units.}
}
\description{
Write dataframe with image url to excel file with image column
}
