% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{c_datatable}
\alias{c_datatable}
\title{Set defaults for function
Create an HTML table widget using the DataTables library with Chinese menu}
\usage{
c_datatable(...)
}
\arguments{
\item{...}{Arguments to \code{DT::datatable}}
}
\description{
Set defaults for function
Create an HTML table widget using the DataTables library with Chinese menu
}
